requirements/base.txt
# Core
Django==5.1.5
django-configurations==2.5.1
django-cors-headers==4.6.0
gunicorn==23.0.0
newrelic==10.3.1
setuptools==75.6.0
python-dotenv==1.0.1

# For the persistence stores
psycopg2-binary==2.9.10
dj-database-url==2.3.0

# Model Tools
django-model-utils==5.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.15.2
django-filter==24.3

# Developer Tools
ipdb==0.13.13
ipython==8.30.0
mkdocs==1.6.1
drf-spectacular==0.28.0

# Testing
mock==5.1.0
factory-boy==3.3.1
coverage==7.6.9

# Static and Media Storage
django-storages==1.14.4
boto3==1.35.76

#auth
djangorestframework-simplejwt==5.4.0
